Refactor(Firebase Import): Saving work in progress
