+++ src/lib/firebase.js
@@ -1,11 +1,7 @@
-// NOTE: import only the Firebase modules that you need in your app... except
-// for the second line, which makes both the linter and react-firebase happy
-import firebase from 'firebase/app';
+import * as firebase from 'firebase/app';
 import 'firebase/firestore';
 
-// Initalize Firebase.
-// These details will need to be replaced with the project specific env vars at the start of each new cohort.
-var firebaseConfig = {
+const firebaseConfig = {
   apiKey: 'AIzaSyDl9f_P7oDF2v5CR-5RHbW7xfysMs4OpP0',
   authDomain: 'tcl-11-smart-shopping-list.firebaseapp.com',
   databaseURL: 'https://tcl-11-smart-shopping-list.firebaseio.com',
@@ -14,7 +10,10 @@ var firebaseConfig = {
   messagingSenderId: '883730532272',
   appId: '1:883730532272:web:45b002254a4bf323273875',
 };
+// Initialize Firebase
+firebase.initializeApp(firebaseConfig);
 
-let fb = firebase.initializeApp(firebaseConfig);
+//To reference database
+export const dataBase = firebase.firestore();
 
-export { fb };
+export default firebase;
